requirements.txt
dotenv
nats-py